refactor: Move table rendering to virtualized components
