+++ src/app/analyze/virtualized-table/virtualized-table-footer.tsx
@@ -3,17 +3,25 @@
 import { TableCell, TableFooter, TableRow } from "@/components/custom/table";
 import * as React from "react";
 import { cn } from "@/lib/utils";
-import { AggregationType } from "../../components/data-table/data-table-aggregations";
+import { AggregationType } from "../../../components/data-table/data-table-aggregations";
 import { useDataTable } from "@/components/data-table/data-table-provider";
+import { useVirtualizer, VirtualItem } from "@tanstack/react-virtual";
 
 export interface DataTableFooterProps<TData> {
+    virtualColumns: VirtualItem[];
+    virtualPadding: {
+      left: number | null;
+      right: number | null;
+    };
 }
 
-export function DataTableFooter<TData>({ }: DataTableFooterProps<TData>) {
+export function DataTableFooter<TData>({
+  virtualColumns,
+  virtualPadding
+ }: DataTableFooterProps<TData>) {
   const { footerAggregations = [], table } = useDataTable();
-
   const columns = table.getVisibleFlatColumns();
-
+  
   const getColumnValues = React.useCallback((columnId: string, isTotal: boolean | undefined): any[] => {
     let rows = [];
     if (isTotal) {
@@ -51,27 +59,45 @@ export function DataTableFooter<TData>({ }: DataTableFooterProps<TData>) {
   return (
     <TableFooter 
       data-testid="table-footer" 
-      className={cn("bg-background sticky bottom-0 z-20")} 
+      className={cn("bg-background sticky bottom-0 z-20 grid")} 
     >
       {footerAggregations.map((aggregation, index) => (
         <TableRow
           data-testid={`footer-row-${aggregation.type}`}
           key={`${aggregation.type}-row`}
-          className={cn("bg-muted/50 hover:bg-muted/50 border-t")}
+          className={cn("bg-muted/50 hover:bg-muted/50 border-t flex")}
         >
-          {columns.map((column, colIndex) => {
+          {virtualPadding.left ? (
+            <td
+              className={cn('virtual-left', 'flex')}
+              style={{
+                width: `${virtualPadding.left}px`,
+              }}
+            />
+          ) : null}
+          
+          {virtualColumns.map(virtualColumn => {
+            const column = columns[virtualColumn.index];
+            if (!column) return null;
+            
             const columnId = column.id;
             const values = getColumnValues(columnId, aggregation.isTotal);
             const content = getAggregation(columnId, aggregation.type, values);
+            const isFirstColumn = virtualColumn.index === 0;
+            
             return (
               <TableCell
                 data-testid={`footer-cell-${aggregation.type}-${columnId}`}
                 key={`${aggregation.type}-${columnId}`}
-                className={cn("font-medium relative select-none truncate border-t border-border")}
+                className={cn("font-medium select-none truncate border-t border-border flex items-center")}
+                style={{
+                  width: `${virtualColumn.size}px`,
+                  height: '100%'
+                }}
               >
                 <div className="flex flex-col gap-1">
                   <div className="flex text-xs text-muted-foreground uppercase">
-                    {colIndex === 0 && aggregation.label}
+                    {isFirstColumn && aggregation.label}
                   </div>
                   <div className="text-muted-foreground">
                     {content}
@@ -80,6 +106,15 @@ export function DataTableFooter<TData>({ }: DataTableFooterProps<TData>) {
               </TableCell>
             );
           })}
+          
+          {virtualPadding.right ? (
+            <td
+              className={cn('virtual-right', 'flex')}
+              style={{
+                width: `${virtualPadding.right}px`,
+              }}
+            />
+          ) : null}
         </TableRow>
       ))}
     </TableFooter>
