+++ src/components/custom/table.tsx
@@ -2,11 +2,11 @@ import * as React from "react";
 
 import { cn } from "@/lib/utils";
 
-interface TableProps extends React.HTMLAttributes<HTMLDivElement> {
+interface TableContainerProps extends React.HTMLAttributes<HTMLDivElement> {
   containerClassName?: string;
 }
 
-const Table = React.forwardRef<HTMLDivElement, TableProps>(
+const TableContainer = React.forwardRef<HTMLDivElement, TableContainerProps>(
   ({ className, containerClassName, onScroll, ...props }, ref) => (
     <div
       className={cn("w-full overflow-auto", containerClassName)}
@@ -15,6 +15,27 @@ const Table = React.forwardRef<HTMLDivElement, TableProps>(
       ref={ref}
     >
       <table
+        className={cn("w-full caption-bottom text-sm", className)}
+        {...props}
+      />
+    </div>
+  )
+);
+TableContainer.displayName = "TableContainer";
+
+interface TableProps extends React.HTMLAttributes<HTMLTableElement> {
+  containerClassName?: string;
+}
+
+const Table = React.forwardRef<HTMLTableElement, TableProps>(
+  ({ className, containerClassName, onScroll, ...props }, ref) => (
+    <div
+      className={cn("w-full overflow-auto", containerClassName)}
+      // REMINDER: we are not scrolling the table, but the container
+      {...{ onScroll }}
+    >
+      <table
+        ref={ref}
         className={cn("w-full caption-bottom text-sm", className)}
         {...props}
       />
@@ -113,6 +134,7 @@ const TableCaption = React.forwardRef<
 TableCaption.displayName = "TableCaption";
 
 export {
+  TableContainer,
   Table,
   TableHeader,
   TableBody,
